commit stable version
