+++ src/aca_main.cpp
@@ -14,9 +14,7 @@
 
 #include "aca_log.h"
 #include "aca_util.h"
-// #include "aca_comm_mgr.h"
-// #include "aca_dataplane_mizar.h"
-#include "messageconsumer.h"
+#include "aca_message_consumer.h"
 #include "aca_async_grpc_server.h"
 #include "goalstateprovisioner.grpc.pb.h"
 #include "cppkafka/utils/consumer_dispatcher.h"
@@ -35,7 +33,6 @@ static char KAFKA_GROUP_ID[] = "test-group-id";
 static char LOCALHOST[] = "localhost";
 static char UDP_PROTOCOL[] = "udp";
 
-// using aca_comm_manager::Aca_Comm_Manager;
 using namespace std;
 
 // Global variables
