+++ src/comm/aca_message_producer.cpp
@@ -14,7 +14,7 @@
 
 #include <iostream>
 #include <stdexcept>
-#include "messageproducer.h"
+#include "aca_message_producer.h"
 
 using namespace std;
 using namespace cppkafka;
